+++ src/tests/test_typeconversions.py
@@ -1142,12 +1142,6 @@ def test_to_py(selenium):
         )
 
 
-def test_pyimport_deprecation(selenium):
-    selenium.run_js("pyodide.runPython('x = 1')")
-    assert selenium.run_js("return pyodide.pyimport('x') === 1")
-    assert "pyodide.pyimport is deprecated and will be removed" in selenium.logs
-
-
 def test_buffer_format_string(selenium):
     errors = [
         ["aaa", "Expected format string to have length <= 2, got 'aaa'"],